tidy up some int casts
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1861176 13f79535-47bb-0310-9956-ffa450edef68

+++ src/scratchpad/src/org/apache/poi/hdgf/chunks/ChunkHeader.java
@@ -63,7 +63,7 @@ public abstract class ChunkHeader {
 			ch.setUnknown2(LittleEndian.getUByte(data, offset + 4));
 			ch.setUnknown3(LittleEndian.getUByte(data, offset + 5));
 			ch.setUnknown1(LittleEndian.getShort(data, offset + 6));
-			ch.setLength((int)LittleEndian.getUInt(data, offset + 8));
+			ch.setLength(Math.toIntExact(LittleEndian.getUInt(data, offset + 8)));
 
 			return ch;
 		} else {
